updated qfs configuration
Updated the QFS configuration by consolidating its configurations to spark-defaults.conf
and make 2x replication the default for files created on QFS. Also make the build-images.sh
script able to take arguments to be passed to the docker file builds, and made the
worker-node Dockerfile configurable through ARG parameters. The intent here is to be able
to build the images with different versions of key software.