updated simple spark cluster configuration
Update the docker configuration of the simple spark cluster to bring it more inline with
the learnings I've made setting up the Spark-QFS cluster. This cluster depends on the
stack being deployed with all nodes mounting a file system, such os GlusterFS or NFS.

The main change over the previous version of the simple Spark cluster is that the
Spark node docker image is built in this build rather than using a third party image. Also
improved the deployment of the Jupyter server.